HCAP-1083 | Backend Developmet ---- - [x] Create Migration - [x] Refatcor validation.js
