+++ server/constants/index.js
@@ -1,0 +1,7 @@
+const ros = require('./ros');
+const validationConstants = require('./validation-constants');
+
+module.exports = {
+  ...validationConstants,
+  ...ros,
+};
